+++ app/Http/Controllers/Examples/SeniorityEnum.php
@@ -1,0 +1,17 @@
+<?php
+//please ignore the namespace ;)
+
+namespace App\Http\Controllers\Examples;
+
+use LaravelEnso\Helpers\app\Classes\Enum;
+
+class SeniorityEnum extends Enum
+{
+    protected static $data = [
+        1 => 'Assistant',
+        2 => 'Associate',
+        3 => 'Staff Member',
+        4 => 'Senior',
+        5 => 'Partner',
+    ];
+}